backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
python-dotenv==1.0.0
selenium==4.16.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
openai==1.6.1
chromadb==0.4.22
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10
numpy==1.26.3
pandas==2.1.4
celery==5.3.4
redis==5.0.1
Pillow==10.2.0
requests==2.31.0
